Commit all fixes
