+++ programs/ralli-bet/programs/ralli-bet/src/instructions/mod.rs
@@ -5,7 +5,7 @@ pub mod join_game;
 // pub mod resolve_game;
 pub mod cancel_game;
 // // pub mod update_game;
-pub mod refund_entry;
+pub mod withdraw_submission;
 
 pub use create_game::*;
 pub use join_game::*;
@@ -15,5 +15,4 @@ pub use join_game::*;
 pub use cancel_game::*;
 // // pub use update_game::*;
 
-pub use refund_entry::*;
-pub use refund_entry::*;
+pub use withdraw_submission::*;
